Disable REST by default, and emit a warning if enabled
After todays meeting, I am hesitant to enable the rest endpoint by default.

Usually, to gain access to data you need access to a database, knowledge of the type of
database, and credentials for it.

With this, there are no credentials, and it runs on a port that you already have open.

This change switches back to being disabled by default, and adds a warning log if it's enabled at application
startup.